Person records introduced
book_ingest_create_records allows now extracting new persons from the metadata and creating a book record out of it.
From the person sections of gndparse some mistakes were removed, it is now possible to check multiple records in the database (as found when there is no ID for a search) if they are of the correct type of person.
The function of adding a record by entering its GND number also checks if it is not by mistake already in the database. main.py and classes were changed a bit to allow for these new modules.