Introduction of Dates_Parsing
The first version of a programme to parse information on lifespans from the GND (dates_parsing.py) has been integrated into the system. It is operated from book_ingest_create_record.py, and some minor changes in books_parsing_bibliographies and classes.py were necessary. Classes also contains the first new draft of a class Making_process that is not yet in use.