person_relation added
The module person_relation was added that contains two functions. The first, already in use, renders an indication of interersonal relations from the GND to the standard used for the database, book_ingest_create_records was adopted to deal with it, and an older error with ID numbers was removed in one case (it may still be extant in another case). The second module, providing inverted relations, should be ready but has not yet been integrated.