All connections established
Now book_ingest_create_records has all modules to create connections between persons, places, and locations, the commands in dbactions have been modified to allow for that (and to allow for changing multiple elements in an 'coonnected person' etc. array (the hardest bit). person_relations now has nine modules, for all possible connections between persons, organisations, and places, and a unified module that creates the term for the reverse connection.