Making Process completed
The selected Making_Process information is now  copied from metadata to pages (originally it was only artist's names). This meant changes in book_ingest_create_records 695-737 + in 494-506, in 481, a mistake was corrected. In this context in gndparse, line 402 was added, in classes line 533 changed. In dbactions (ine 88)and in main (line 265), the printout of the entire resources list was commented out because I found it annoying, in display-record a function that had been commented out was restored to allow for the display of persons etc.