First stage of Refactoring
parse_gnd: major refactoring for hopefully all the parsing of the XML records.
However: a number of functions are commented out because Entity currently does not allow attaching the results:
The search actions in identify_persons etc. were not touched since they need changes because of the simpler data structure. Currently, identify_organisations and identify_places do not seem to function, and hence the related parsing could not be tested at all (I need either them working or a functioning test routine for the individual functions).
In parse_istc some functions, especially the date, were refactored, in parse_iiif there was only a minor correction of a horrible loop.