Refactoring parse_vd17_vd18.py
The parsing for boks from the VD17 and VD18 has been removed from books_parsing_bibliographies, placed in parse_vd17_vd18, and refactored.

There is one greater problem: The VD17/18 give GND IDs for authors, printers etc., but since they are not given by the ISTC, there is currently no place for them in the structure of the record. I didn't change it because the role, eac etc records will be radically changed, anyway.
In the other files, there were hardly any changes.