beanie_text4 added otherwise only small changes
beanie_text4 is for deleting.

I didn't get the deletion of connected records work, hence I delete the Entity, EntityConnections, and EntityAndConnections separately.
Problem: I have to remove the link to EntityConnection from all other EntityAndConnection records of which it is part. This seems to work, but when doing so it seems to duplicate the EntityConnection records and the connected Entity records that are linked to this EntityAndConnections record and have not been deleted - why???