New Tests 5 and6 (wsith classes and dbactions adjusted)
Test 5 experiments with a Bridge class that consists of two Connection instances. The mixture between array in one document and array in links apparently confuses Beanie, and I did not find a simple way to querying in.
I then made my wayto Delphi, and the Pythia suggested a complicated pipeline that, however, does not lead to complete results (and I have doubts about scalability).

test6 takes it further, here, the Bridge instance is merely linked to two (in theory any number, in reality always two) Coinnection instances.
Here, the search is easy - for display purposes I created something like the old EntityAndConnections as agreed yesterday.

To me, the structure of 6 seems fine  - but maybe I overlooked some bad details. This was "CR" - I hope to try "UD" in the next days - however, since the data structure is now simpler, this shoudl work well.